refactor(autonomi): simplify external signer put
