+++ autonomi/src/lib.rs
@@ -35,11 +35,14 @@ extern crate tracing;
 pub mod client;
 #[cfg(feature = "data")]
 mod self_encryption;
+mod utils;
 
 pub use sn_evm::get_evm_network_from_env;
 pub use sn_evm::EvmNetwork;
 pub use sn_evm::EvmWallet as Wallet;
 pub use sn_evm::RewardsAddress;
+#[cfg(feature = "external-signer")]
+pub use utils::payment_proof_from_quotes_and_payments;
 
 #[doc(no_inline)] // Place this under 'Re-exports' in the docs.
 pub use bytes::Bytes;
